refactor: move and rename ResourceSelectorCard
